frontend(types): reduce deterministic lint violations in packages/types
Focused fixes for jsdoc/max-lines/id-length/no-magic-numbers in packages/types with real code changes only.

- Split types.ts to satisfy max-lines and keep API via re-exports
- Added typed JSDoc tags for utility functions
- Replaced magic numbers with named constants
- Renamed short identifier in velocity reducer
